refactor(cloud): extract hosted route views
