+++ apps/notebook-cloud/viewer/cloud-route-views.tsx
@@ -1,0 +1,805 @@
+import { useCallback, useEffect, useMemo, useState, type FormEvent } from "react";
+import {
+  AlertCircle,
+  BookOpen,
+  FilePlus2,
+  KeyRound,
+  Loader2,
+  LogIn,
+  LogOut,
+  RotateCcw,
+  UserRound,
+} from "lucide-react";
+import { useTheme } from "@/hooks/useTheme";
+import {
+  clearCloudPrototypeDevAuth,
+  fetchWithCloudPrototypeAuth,
+  prepareCloudOidcViewerLogin,
+  type CloudPrototypeAuthState,
+} from "./collaborator-auth";
+import { beginOidcLogin, completeOidcRedirect } from "./oidc-auth";
+import { cloudResponseError } from "./cloud-response";
+import { clearCloudAppSession, establishCloudAppSessionFromOidcToken } from "./app-session";
+import { CloudNotebookDashboard } from "./cloud-notebook-dashboard-view";
+import { isNotebookListPath, loadCloudNotebookListBootstrap } from "./cloud-viewer-config";
+import type {
+  CloudNotebookCreateResponse,
+  CloudNotebookListBootstrap,
+  CloudNotebookListResponse,
+  CloudNotebookListState,
+  CloudNotebookRenameState,
+  CloudNotebookUpdateResponse,
+  CloudViewerAuthConfig,
+} from "./cloud-viewer-types";
+import type { ViewerStatus } from "./notice-types";
+import { projectCloudNotebookDashboard, type CloudNotebookListItem } from "./notebook-dashboard";
+import {
+  clearCachedCloudNotebookList,
+  readCachedCloudNotebookList,
+  writeCachedCloudNotebookList,
+} from "./notebook-list-cache";
+import { applyDocumentTheme, CLOUD_VIEWER_THEME_STORAGE_KEY } from "./theme";
+import {
+  useCloudAppSessionBridge,
+  useCloudAppSessionStatus,
+  useCloudPrototypeAuth,
+} from "./use-cloud-auth";
+import { CloudNotebookSignInButton } from "./cloud-auth-controls";
+
+export function CloudNotebookListView({ authConfig }: { authConfig: CloudViewerAuthConfig }) {
+  const { resolvedTheme } = useTheme(CLOUD_VIEWER_THEME_STORAGE_KEY);
+  const [bootstrap, setBootstrap] = useState<CloudNotebookListBootstrap | null>(() =>
+    loadCloudNotebookListBootstrap(),
+  );
+  const appSessionStatus = useCloudAppSessionStatus(bootstrap?.session ?? null);
+  const { authState, authRenewal, refreshAuthState } = useCloudPrototypeAuth(authConfig, {
+    appSessionRefreshFallback: true,
+    appSessionLoading: appSessionStatus.status === "loading",
+    appSession: appSessionStatus.session,
+  });
+  useCloudAppSessionBridge(
+    authState,
+    appSessionStatus.session,
+    appSessionStatus.status === "loading",
+    appSessionStatus.refreshAppSessionStatus,
+  );
+  const [listState, setListState] = useState<CloudNotebookListState>(() =>
+    initialCloudNotebookListState(authState, bootstrap),
+  );
+  const [refreshIndex, setRefreshIndex] = useState(0);
+  const [createState, setCreateState] = useState<"idle" | "starting">("idle");
+  const [createError, setCreateError] = useState<string | null>(null);
+  const [createFormOpen, setCreateFormOpen] = useState(false);
+  const [createTitle, setCreateTitle] = useState(() => defaultCloudNotebookTitle());
+  const [renameState, setRenameState] = useState<CloudNotebookRenameState | null>(null);
+  const [renameSavingId, setRenameSavingId] = useState<string | null>(null);
+  const [renameError, setRenameError] = useState<string | null>(null);
+  const hasExplicitAuth = authState.mode === "dev" || authState.mode === "oidc";
+  const hasAppSession = Boolean(appSessionStatus.session);
+  const signedIn = hasExplicitAuth || hasAppSession;
+  const canFetchNotebookList = authState.mode === "dev" || hasAppSession;
+  const waitingForAppSession = authState.mode === "oidc" && !hasAppSession;
+  const dashboardModel = useMemo(
+    () => (listState.kind === "ready" ? projectCloudNotebookDashboard(listState.notebooks) : null),
+    [listState],
+  );
+
+  useEffect(() => {
+    applyDocumentTheme(resolvedTheme);
+  }, [resolvedTheme]);
+
+  useEffect(() => {
+    const cachedNotebooks =
+      readCachedCloudNotebookListFromWindow(authState) ?? bootstrap?.notebooks ?? null;
+    if (!canFetchNotebookList) {
+      if (waitingForAppSession) {
+        setListState(
+          cachedNotebooks ? { kind: "ready", notebooks: cachedNotebooks } : { kind: "loading" },
+        );
+        return;
+      }
+      clearCachedCloudNotebookListFromWindow();
+      setListState({ kind: "signed_out" });
+      return;
+    }
+
+    const controller = new AbortController();
+    setListState(
+      cachedNotebooks ? { kind: "ready", notebooks: cachedNotebooks } : { kind: "loading" },
+    );
+    void (async () => {
+      try {
+        const response = await fetchCloudNotebookList(authState, controller.signal);
+        if (controller.signal.aborted) return;
+        if (!response.ok) {
+          throw await cloudResponseError(response, "Unable to list notebooks");
+        }
+        const body = (await response.json()) as unknown;
+        if (!isCloudNotebookListResponse(body)) {
+          throw new Error("Unable to list notebooks: response shape was invalid");
+        }
+        writeCachedCloudNotebookListToWindow(authState, body.notebooks);
+        setListState({ kind: "ready", notebooks: body.notebooks });
+      } catch (error) {
+        if (controller.signal.aborted) return;
+        setListState({
+          kind: "error",
+          message: error instanceof Error ? error.message : String(error),
+        });
+      }
+    })();
+
+    return () => {
+      controller.abort();
+    };
+  }, [authState, bootstrap, canFetchNotebookList, refreshIndex, waitingForAppSession]);
+
+  const refreshList = () => {
+    setRefreshIndex((value) => value + 1);
+  };
+
+  const openCreateForm = () => {
+    if (!signedIn) {
+      return;
+    }
+    setCreateError(null);
+    setCreateTitle(defaultCloudNotebookTitle());
+    setCreateFormOpen(true);
+  };
+
+  const closeCreateForm = () => {
+    if (createState === "starting") {
+      return;
+    }
+    setCreateError(null);
+    setCreateFormOpen(false);
+  };
+
+  const createNotebook = async (event: FormEvent<HTMLFormElement>) => {
+    event.preventDefault();
+    if (!signedIn || createState === "starting") {
+      return;
+    }
+    const title = createTitle.trim() || defaultCloudNotebookTitle();
+    try {
+      setCreateError(null);
+      setCreateState("starting");
+      const response = await fetchWithCloudPrototypeAuth(
+        cloudNotebookCollectionEndpoint(),
+        {
+          method: "POST",
+          headers: {
+            Accept: "application/json",
+            "Content-Type": "application/json",
+          },
+          body: JSON.stringify({ title }),
+        },
+        cloudAuthWithScope(authState, "owner"),
+      );
+      if (!response.ok) {
+        throw await cloudResponseError(response, "Unable to create notebook");
+      }
+      const body = (await response.json()) as CloudNotebookCreateResponse;
+      if (body.ok !== true || typeof body.viewer_url !== "string") {
+        throw new Error("Unable to create notebook: response shape was invalid");
+      }
+      window.location.assign(body.viewer_url);
+    } catch (error) {
+      setCreateState("idle");
+      setCreateError(error instanceof Error ? error.message : String(error));
+    }
+  };
+
+  const openRenameForm = useCallback((notebook: CloudNotebookListItem) => {
+    setRenameError(null);
+    setRenameState({
+      notebookId: notebook.notebook_id,
+      title: notebook.title?.trim() ?? "",
+    });
+  }, []);
+
+  const closeRenameForm = useCallback(() => {
+    if (renameSavingId) {
+      return;
+    }
+    setRenameError(null);
+    setRenameState(null);
+  }, [renameSavingId]);
+
+  const saveNotebookTitle = async (event: FormEvent<HTMLFormElement>) => {
+    event.preventDefault();
+    if (!signedIn || !renameState || renameSavingId) {
+      return;
+    }
+
+    const notebookId = renameState.notebookId;
+    const nextTitle = renameState.title.trim();
+    try {
+      setRenameError(null);
+      setRenameSavingId(notebookId);
+      const response = await fetchWithCloudPrototypeAuth(
+        cloudNotebookCatalogEndpoint(notebookId),
+        {
+          method: "PATCH",
+          headers: {
+            Accept: "application/json",
+            "Content-Type": "application/json",
+          },
+          body: JSON.stringify({ title: nextTitle || null }),
+        },
+        cloudAuthWithScope(authState, "editor"),
+      );
+      if (!response.ok) {
+        throw await cloudResponseError(response, "Unable to rename notebook");
+      }
+      const body = (await response.json()) as CloudNotebookUpdateResponse;
+      if (body.ok !== true || body.notebook_id !== notebookId) {
+        throw new Error("Unable to rename notebook: response shape was invalid");
+      }
+      setListState((current) => {
+        if (current.kind !== "ready") {
+          return current;
+        }
+        const notebooks = current.notebooks.map((notebook) =>
+          notebook.notebook_id === notebookId
+            ? {
+                ...notebook,
+                title: body.title ?? null,
+                updated_at: body.updated_at ?? notebook.updated_at,
+                viewer_url: body.viewer_url ?? notebook.viewer_url,
+              }
+            : notebook,
+        );
+        writeCachedCloudNotebookListToWindow(authState, notebooks);
+        return {
+          kind: "ready",
+          notebooks,
+        };
+      });
+      setRenameState(null);
+    } catch (error) {
+      setRenameError(error instanceof Error ? error.message : String(error));
+    } finally {
+      setRenameSavingId(null);
+    }
+  };
+
+  const signOut = () => {
+    setBootstrap(null);
+    appSessionStatus.clearAppSessionStatus();
+    clearCachedCloudNotebookListFromWindow();
+    void clearCloudAppSession()
+      .catch((error: unknown) => {
+        console.warn("[notebook-cloud] app session clear failed", error);
+      })
+      .finally(appSessionStatus.refreshAppSessionStatus);
+    clearCloudPrototypeDevAuth(window.localStorage);
+    refreshAuthState();
+  };
+
+  const headerDetail = cloudNotebookListHeaderDetail(authState, hasAppSession);
+
+  return (
+    <main className="cloud-notebook-list-page">
+      <header className="cloud-notebook-list-header">
+        <div>
+          <a className="cloud-notebook-list-brand" href="/">
+            nteract
+          </a>
+          <h1>Notebooks</h1>
+          <p>{headerDetail}</p>
+        </div>
+        <div className="cloud-notebook-list-actions">
+          <button
+            type="button"
+            disabled={!signedIn || listState.kind === "loading"}
+            onClick={refreshList}
+          >
+            <RotateCcw aria-hidden="true" />
+            Refresh
+          </button>
+          <button
+            type="button"
+            disabled={!signedIn || createState === "starting"}
+            onClick={openCreateForm}
+          >
+            {createState === "starting" ? (
+              <Loader2 className="cloud-home-status-spinner" aria-hidden="true" />
+            ) : (
+              <FilePlus2 aria-hidden="true" />
+            )}
+            {createState === "starting" ? "Creating" : "New notebook"}
+          </button>
+          {hasExplicitAuth || hasAppSession ? null : (
+            <CloudNotebookSignInButton authConfig={authConfig} authState={authState} />
+          )}
+          {signedIn ? (
+            <button type="button" onClick={signOut}>
+              <LogOut aria-hidden="true" />
+              Sign out
+            </button>
+          ) : null}
+        </div>
+      </header>
+
+      {authRenewal.kind !== "idle" && !hasAppSession ? (
+        <div
+          className="cloud-notebook-list-banner"
+          data-kind={authRenewal.kind === "failed" ? "error" : "info"}
+          role={authRenewal.kind === "failed" ? "alert" : "status"}
+        >
+          {authRenewal.message}
+        </div>
+      ) : null}
+      {createError ? (
+        <div className="cloud-notebook-list-banner" data-kind="error" role="alert">
+          {createError}
+        </div>
+      ) : null}
+      {renameError ? (
+        <div className="cloud-notebook-list-banner" data-kind="error" role="alert">
+          {renameError}
+        </div>
+      ) : null}
+      {createFormOpen ? (
+        <form className="cloud-new-notebook-form" onSubmit={createNotebook}>
+          <label htmlFor="cloud-new-notebook-title">Notebook title</label>
+          <input
+            id="cloud-new-notebook-title"
+            type="text"
+            value={createTitle}
+            maxLength={160}
+            disabled={createState === "starting"}
+            onChange={(event) => setCreateTitle(event.currentTarget.value)}
+          />
+          <button type="submit" disabled={createState === "starting"}>
+            {createState === "starting" ? (
+              <Loader2 className="cloud-home-status-spinner" aria-hidden="true" />
+            ) : (
+              <FilePlus2 aria-hidden="true" />
+            )}
+            Create
+          </button>
+          <button type="button" disabled={createState === "starting"} onClick={closeCreateForm}>
+            Cancel
+          </button>
+        </form>
+      ) : null}
+
+      <section className="cloud-notebook-list-content" aria-label="Notebook list">
+        {listState.kind === "loading" ? (
+          <div className="cloud-notebook-list-state" data-kind="loading" role="status">
+            <Loader2 className="cloud-home-status-spinner" aria-hidden="true" />
+            <span>Loading notebooks</span>
+          </div>
+        ) : listState.kind === "signed_out" ? (
+          <div className="cloud-notebook-list-state" data-kind="signed-out">
+            <KeyRound aria-hidden="true" />
+            <span>Sign in to view notebooks.</span>
+          </div>
+        ) : listState.kind === "error" ? (
+          <div className="cloud-notebook-list-state" data-kind="error" role="alert">
+            <AlertCircle aria-hidden="true" />
+            <span>{listState.message}</span>
+          </div>
+        ) : listState.notebooks.length === 0 ? (
+          <div className="cloud-notebook-list-state" data-kind="empty">
+            <BookOpen aria-hidden="true" />
+            <span>No notebooks yet.</span>
+          </div>
+        ) : dashboardModel ? (
+          <CloudNotebookDashboard
+            model={dashboardModel}
+            canRename={signedIn}
+            renameState={renameState}
+            renameSavingId={renameSavingId}
+            onOpenRename={openRenameForm}
+            onCancelRename={closeRenameForm}
+            onRenameTitleChange={(title) =>
+              setRenameState((current) => (current ? { ...current, title } : current))
+            }
+            onSaveRename={saveNotebookTitle}
+          />
+        ) : (
+          <div className="cloud-notebook-list-state" data-kind="error" role="alert">
+            <AlertCircle aria-hidden="true" />
+            <span>Unable to project notebook dashboard.</span>
+          </div>
+        )}
+      </section>
+    </main>
+  );
+}
+
+function cloudNotebookListHeaderDetail(
+  authState: CloudPrototypeAuthState,
+  hasAppSession: boolean,
+): string {
+  if (authState.mode === "oidc_expired") {
+    return "Session expired";
+  }
+  if (authState.mode === "anonymous" && !hasAppSession) {
+    return "Signed out";
+  }
+  const firstName = cloudNotebookListFirstName(authState);
+  return firstName ? `by ${firstName}` : "Signed in";
+}
+
+function cloudNotebookListFirstName(authState: CloudPrototypeAuthState): string | null {
+  const claimName =
+    authState.oidcClaims?.given_name?.trim() || authState.oidcClaims?.name?.trim() || "";
+  if (!claimName || claimName.includes("@")) {
+    return null;
+  }
+  return claimName.split(/\s+/u)[0] ?? null;
+}
+
+function cloudNotebookListEndpoint(): string {
+  return new URL("api/n?limit=100", `${window.location.origin}/`).href;
+}
+
+function cloudNotebookCollectionEndpoint(): string {
+  return new URL("api/n", `${window.location.origin}/`).href;
+}
+
+function cloudNotebookCatalogEndpoint(notebookId: string): string {
+  return new URL(`api/n/${encodeURIComponent(notebookId)}`, `${window.location.origin}/`).href;
+}
+
+function cloudAuthWithScope(
+  authState: CloudPrototypeAuthState,
+  requestedScope: NonNullable<CloudPrototypeAuthState["requestedScope"]>,
+): CloudPrototypeAuthState {
+  return authState.requestedScope === requestedScope
+    ? authState
+    : {
+        ...authState,
+        requestedScope,
+      };
+}
+
+function initialCloudNotebookListState(
+  authState: CloudPrototypeAuthState,
+  bootstrap: CloudNotebookListBootstrap | null,
+): CloudNotebookListState {
+  const cachedNotebooks = readCachedCloudNotebookListFromWindow(authState) ?? bootstrap?.notebooks;
+  return cachedNotebooks ? { kind: "ready", notebooks: cachedNotebooks } : { kind: "loading" };
+}
+
+function fetchCloudNotebookList(
+  authState: CloudPrototypeAuthState,
+  signal: AbortSignal,
+): Promise<Response> {
+  if (authState.mode === "dev" || authState.mode === "oidc") {
+    return fetchWithCloudPrototypeAuth(
+      cloudNotebookListEndpoint(),
+      { headers: { Accept: "application/json" }, signal },
+      authState,
+    );
+  }
+  return fetch(cloudNotebookListEndpoint(), {
+    credentials: "same-origin",
+    headers: { Accept: "application/json" },
+    signal,
+  });
+}
+
+function readCachedCloudNotebookListFromWindow(
+  authState: CloudPrototypeAuthState,
+): CloudNotebookListItem[] | null {
+  const storage = cloudNotebookListCacheStorage();
+  return storage ? readCachedCloudNotebookList(storage, authState) : null;
+}
+
+function writeCachedCloudNotebookListToWindow(
+  authState: CloudPrototypeAuthState,
+  notebooks: CloudNotebookListItem[],
+): void {
+  const storage = cloudNotebookListCacheStorage();
+  if (!storage) {
+    return;
+  }
+  writeCachedCloudNotebookList(storage, authState, notebooks);
+}
+
+function clearCachedCloudNotebookListFromWindow(): void {
+  const storage = cloudNotebookListCacheStorage();
+  if (!storage) {
+    return;
+  }
+  clearCachedCloudNotebookList(storage);
+}
+
+function cloudNotebookListCacheStorage(): Storage | null {
+  try {
+    return window.sessionStorage;
+  } catch {
+    return null;
+  }
+}
+
+function defaultCloudNotebookTitle(now = new Date()): string {
+  const date = new Intl.DateTimeFormat(undefined, {
+    month: "short",
+    day: "numeric",
+    year: "numeric",
+  }).format(now);
+  const time = new Intl.DateTimeFormat(undefined, {
+    hour: "numeric",
+    minute: "2-digit",
+  }).format(now);
+  return `Notebook ${date} ${time}`;
+}
+
+function isCloudNotebookListResponse(value: unknown): value is CloudNotebookListResponse {
+  if (!value || typeof value !== "object") {
+    return false;
+  }
+  const candidate = value as Record<string, unknown>;
+  return candidate.ok === true && Array.isArray(candidate.notebooks);
+}
+
+export function CloudHomeView({ authConfig }: { authConfig: CloudViewerAuthConfig }) {
+  const { resolvedTheme } = useTheme(CLOUD_VIEWER_THEME_STORAGE_KEY);
+  const appSessionStatus = useCloudAppSessionStatus(null);
+  const { authState, authRenewal, refreshAuthState } = useCloudPrototypeAuth(authConfig, {
+    appSessionRefreshFallback: true,
+    appSessionLoading: appSessionStatus.status === "loading",
+    appSession: appSessionStatus.session,
+  });
+  useCloudAppSessionBridge(
+    authState,
+    appSessionStatus.session,
+    appSessionStatus.status === "loading",
+    appSessionStatus.refreshAppSessionStatus,
+  );
+  const [authAction, setAuthAction] = useState<"idle" | "starting">("idle");
+  const [formError, setFormError] = useState<string | null>(null);
+  const oidcConfigured = Boolean(authConfig.oidc);
+
+  useEffect(() => {
+    applyDocumentTheme(resolvedTheme);
+  }, [resolvedTheme]);
+
+  const beginOidcAuth = async () => {
+    if (!authConfig.oidc) {
+      setFormError("OIDC sign-in is not configured for this host.");
+      return;
+    }
+    try {
+      setAuthAction("starting");
+      prepareCloudOidcViewerLogin(window.localStorage);
+      const url = await beginOidcLogin(authConfig.oidc, {
+        currentUrl: window.location.href,
+        storage: window.localStorage,
+      });
+      window.location.assign(url.href);
+    } catch (error) {
+      setAuthAction("idle");
+      setFormError(error instanceof Error ? error.message : String(error));
+    }
+  };
+
+  const resetAuth = () => {
+    appSessionStatus.clearAppSessionStatus();
+    void clearCloudAppSession()
+      .catch((error: unknown) => {
+        console.warn("[notebook-cloud] app session clear failed", error);
+      })
+      .finally(appSessionStatus.refreshAppSessionStatus);
+    clearCloudPrototypeDevAuth(window.localStorage);
+    setFormError(null);
+    refreshAuthState();
+  };
+
+  const hasExplicitAuth = authState.mode === "oidc";
+  const hasAppSession = Boolean(appSessionStatus.session);
+  const signedIn = hasExplicitAuth || hasAppSession;
+  const homeStatusMode = signedIn ? "oidc" : authState.mode;
+  const homeStatusTitle = hasExplicitAuth
+    ? (authState.user ?? "Signed in")
+    : hasAppSession
+      ? "Signed in"
+      : "Open a notebook";
+  const homeStatusDescription = signedIn
+    ? hasExplicitAuth
+      ? "Open a notebook or sign out of this browser session."
+      : "Open and manage notebooks with this browser session."
+    : "Sign in to open private notebooks or request edit access.";
+
+  return (
+    <main className="cloud-home">
+      <section className="cloud-home-layout" aria-label="nteract notebook entry">
+        <div className="cloud-home-copy">
+          <h1>nteract</h1>
+          <span>realtime notebooks</span>
+        </div>
+
+        <section
+          className="cloud-home-panel"
+          data-mode={homeStatusMode}
+          aria-label="Notebook sign-in"
+        >
+          <div className="cloud-home-status" data-mode={homeStatusMode}>
+            {signedIn ? <UserRound aria-hidden="true" /> : <KeyRound aria-hidden="true" />}
+            <div>
+              <h2>{homeStatusTitle}</h2>
+              <p>{homeStatusDescription}</p>
+            </div>
+          </div>
+
+          {formError ? (
+            <div className="cloud-auth-form-error" role="alert">
+              {formError}
+            </div>
+          ) : null}
+          {authRenewal.kind !== "idle" && !hasAppSession ? (
+            <div
+              className="cloud-auth-form-error"
+              data-kind={authRenewal.kind === "failed" ? "error" : "info"}
+              role={authRenewal.kind === "failed" ? "alert" : "status"}
+            >
+              {authRenewal.message}
+            </div>
+          ) : null}
+
+          <div className="cloud-home-actions">
+            <a href="/n">View notebooks</a>
+            {signedIn ? (
+              <button
+                type="button"
+                onClick={() => {
+                  appSessionStatus.clearAppSessionStatus();
+                  void clearCloudAppSession()
+                    .catch((error: unknown) => {
+                      console.warn("[notebook-cloud] app session clear failed", error);
+                    })
+                    .finally(appSessionStatus.refreshAppSessionStatus);
+                  clearCloudPrototypeDevAuth(window.localStorage);
+                  refreshAuthState();
+                }}
+              >
+                <LogOut aria-hidden="true" />
+                Sign out
+              </button>
+            ) : null}
+            {hasExplicitAuth ? null : (
+              <button
+                type="button"
+                disabled={authAction === "starting" || !oidcConfigured}
+                onClick={beginOidcAuth}
+              >
+                <LogIn aria-hidden="true" />
+                {authAction === "starting"
+                  ? "Starting sign-in"
+                  : !oidcConfigured
+                    ? "Sign-in unavailable"
+                    : hasAppSession
+                      ? "Renew sign-in"
+                      : "Sign in with Anaconda"}
+              </button>
+            )}
+            {authState.mode === "invalid" || authState.mode === "oidc_expired" ? (
+              <button type="button" onClick={resetAuth}>
+                <RotateCcw aria-hidden="true" />
+                Reset
+              </button>
+            ) : null}
+          </div>
+
+          {oidcConfigured ? null : (
+            <p className="cloud-home-note">
+              This host has no sign-in provider configured. Public notebooks can still be read.
+            </p>
+          )}
+        </section>
+      </section>
+    </main>
+  );
+}
+
+export function OidcCallbackView({ authConfig }: { authConfig: CloudViewerAuthConfig }) {
+  const { resolvedTheme } = useTheme(CLOUD_VIEWER_THEME_STORAGE_KEY);
+  const [status, setStatus] = useState<ViewerStatus>({
+    kind: "loading",
+    message: "Completing sign-in...",
+  });
+
+  useEffect(() => {
+    applyDocumentTheme(resolvedTheme);
+  }, [resolvedTheme]);
+
+  useEffect(() => {
+    const oidc = authConfig.oidc;
+    if (!oidc) {
+      setStatus({ kind: "error", message: "OIDC sign-in is not configured for this host." });
+      return;
+    }
+
+    const params = new URLSearchParams(window.location.search);
+    if (!params.has("code") || !params.has("state")) {
+      setStatus({ kind: "empty", message: "No sign-in callback is pending." });
+      return;
+    }
+
+    let cancelled = false;
+    void completeOidcRedirect(oidc, {
+      callbackUrl: window.location.href,
+      storage: window.localStorage,
+    })
+      .then(async ({ returnUrl, token }) => {
+        if (cancelled) return;
+        await establishCloudAppSessionFromOidcToken(token).catch((error: unknown) => {
+          console.warn("[notebook-cloud] app session exchange failed", error);
+        });
+        if (cancelled) return;
+        setStatus({ kind: "ready", message: "Signed in. Returning to the notebook..." });
+        window.location.replace(returnUrl);
+      })
+      .catch((error: unknown) => {
+        if (cancelled) return;
+        setStatus({
+          kind: "error",
+          message: error instanceof Error ? error.message : String(error),
+        });
+      });
+
+    return () => {
+      cancelled = true;
+    };
+  }, [authConfig.oidc]);
+
+  const statusTitle =
+    status.kind === "ready"
+      ? "Signed in"
+      : status.kind === "error"
+        ? "Sign-in needs attention"
+        : status.kind === "empty"
+          ? "Nothing to finish"
+          : "Completing sign-in";
+  const statusIcon =
+    status.kind === "error" ? (
+      <AlertCircle aria-hidden="true" />
+    ) : status.kind === "ready" ? (
+      <UserRound aria-hidden="true" />
+    ) : status.kind === "empty" ? (
+      <KeyRound aria-hidden="true" />
+    ) : (
+      <Loader2 className="cloud-home-status-spinner" aria-hidden="true" />
+    );
+
+  return (
+    <main className="cloud-home">
+      <section className="cloud-home-layout" aria-label="nteract sign-in callback">
+        <div className="cloud-home-copy">
+          <h1>nteract</h1>
+          <span>returning to the notebook</span>
+        </div>
+
+        <section
+          className="cloud-home-panel"
+          data-mode={status.kind}
+          aria-label="Cloud sign-in status"
+        >
+          <div className="cloud-home-status" data-mode={status.kind}>
+            {statusIcon}
+            <div>
+              <h2>{statusTitle}</h2>
+              <p>{status.message}</p>
+            </div>
+          </div>
+
+          {status.kind === "error" || status.kind === "empty" ? (
+            <div className="cloud-home-actions">
+              <a href="/">Back to nteract</a>
+            </div>
+          ) : null}
+        </section>
+      </section>
+    </main>
+  );
+}
+
+export function shouldBootstrapNotebookListRoute(): boolean {
+  return isNotebookListPath();
+}
